Initial Commit
This program is a recreation of Trojan.Win32.DesktopPuzzle exept it
really does block you from using task manager to close it and there also
are no numbers to mark the order of the tiles. To move a tile just click
it. TO CLOSE: Push the ESC key or complete the puzzle.